pyproject.toml
[project]
name = "a4i"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.16.0",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "ag-ui-adk>=0.3.1",
    "googlemaps>=4.10.0",
]

[tool.setuptools]
packages = [
    "first_responder_agent",
    "first_responder_agent.common",
    "first_responder_agent.disaster_discovery_agent",
    "first_responder_agent.disaster_discovery_agent.fema_live_agent",
    "first_responder_agent.disaster_discovery_agent.noaa_live_agent",
    "first_responder_agent.relief_finder_agent",
    "first_responder_agent.insights_agent",
]
